Updated Project Structure
Moved DataBaseLogin into ui-Operations
